requirements.txt
bokeh==2.2.0
altair==4.1.0
scipy==1.6.2
streamlit==0.79.0
plotly==4.5.0
numpy==1.20.2
matplotlib==3.4.0
pandas==1.2.3
Pillow==8.1.2
ngboost==0.3.10
scikit-learn==0.23.2
